Breaks the extension methods out into more relevantly named files.
